Commit 2 (atual)

+++ repository/employeeRepository.php
@@ -1,5 +1,9 @@
 
 <?php
+
+use Illuminate\Database\Capsule\Manager as Capsule;
+use Labsoft\Models\Employee;
+
 class EmployeeRepository{
 
     private $mySql;
@@ -11,6 +15,24 @@ class EmployeeRepository{
     public function findAll(){
 
         try{
+            if ($this->canUseEloquent()) {
+                return Capsule::table('employee as e')
+                    ->select([
+                        'e.id',
+                        'e.name',
+                        'e.position',
+                        'e.created_date',
+                        'u1.nome as user_name',
+                        'e.last_modified_date',
+                        'u2.nome as last_user_name',
+                    ])
+                    ->leftJoin('usuario as u1', 'e.created_by', '=', 'u1.id')
+                    ->leftJoin('usuario as u2', 'e.last_modified_by', '=', 'u2.id')
+                    ->orderBy('e.name', 'asc')
+                    ->get()
+                    ->toArray();
+            }
+
             $sql = "SELECT e.id, name, position, created_date, 
                         (SELECT nome FROM usuario WHERE id = e.created_by) AS user_name, 
                         last_modified_date, 
@@ -30,6 +52,14 @@ class EmployeeRepository{
     public function findByName($name){
 
         try{
+            if ($this->canUseEloquent()) {
+                return Employee::query()
+                    ->select(['id', 'name', 'position', 'created_date'])
+                    ->where('name', $name)
+                    ->get()
+                    ->toArray();
+            }
+
             $sql = "SELECT id, name, position, created_date
                     FROM employee
                     WHERE name = '".$name."'"; 
@@ -45,6 +75,15 @@ class EmployeeRepository{
 
     public function findByNameAndPosition($name, $position){
         try{
+            if ($this->canUseEloquent()) {
+                return Employee::query()
+                    ->select(['id', 'name', 'position', 'created_date'])
+                    ->where('name', $name)
+                    ->where('position', $position)
+                    ->get()
+                    ->toArray();
+            }
+
             $sql = "SELECT id, name, position, created_date
                     FROM employee
                     WHERE name = '".$name."' 
@@ -62,6 +101,16 @@ class EmployeeRepository{
     public function save($post){
 
         try{
+            if ($this->canUseEloquent()) {
+                Employee::query()->create([
+                    'name' => $post['name'],
+                    'position' => $post['position'],
+                    'created_date' => date('Y-m-d H:i:s'),
+                    'created_by' => $_SESSION['id'],
+                ]);
+                return 'SAVED';
+            }
+
             $sql = "INSERT INTO employee (name, position, created_date, created_by)
                     VALUES(
                      '".$post['name']."',
@@ -82,6 +131,18 @@ class EmployeeRepository{
     public function updateById($id, $name, $position){
 
         try{
+            if ($this->canUseEloquent()) {
+                Employee::query()
+                    ->where('id', (int) $id)
+                    ->update([
+                        'name' => $name,
+                        'position' => $position,
+                        'last_modified_date' => date('Y-m-d H:i:s'),
+                        'last_modified_by' => $_SESSION['id'],
+                    ]);
+                return 'UPDATED';
+            }
+
             $sql = "UPDATE employee
                     SET name = '".$name."', position = '".$position."', last_modified_date = '".date("Y-m-d H:i:s")."', last_modified_by = ".$_SESSION['id']." 
                     WHERE ID = ".$id;  
@@ -97,6 +158,13 @@ class EmployeeRepository{
     public function deleteById($id){
 
         try{
+            if ($this->canUseEloquent()) {
+                Employee::query()
+                    ->where('id', (int) $id)
+                    ->delete();
+                return 'DELETED';
+            }
+
             $sql = "DELETE FROM employee
                     WHERE id = ".$id;  
 
@@ -106,6 +174,10 @@ class EmployeeRepository{
         }catch(Exception $e){
             return 'DELETE_ERROR';
         }
+    }
+
+    private function canUseEloquent(){
+        return class_exists(Employee::class) && class_exists(Capsule::class);
     }
 }
 ?>
