+++ repository/shippingCompanyRepository.php
@@ -1,4 +1,8 @@
 <?php
+
+use Illuminate\Database\Capsule\Manager as Capsule;
+use Labsoft\Models\ShippingCompany;
+
 class ShippingCompanyRepository{
 
     private $mySql;
@@ -10,6 +14,26 @@ class ShippingCompanyRepository{
     public function findByClient($clientName){
 
         try{
+            if ($this->canUseEloquent()) {
+                return ShippingCompany::query()
+                    ->select([
+                        'id',
+                        'nome',
+                        'username',
+                        'cnpj',
+                        'email',
+                        'telefone',
+                        'password',
+                        'data',
+                        'usuario',
+                        'celular',
+                        'cliente_origem',
+                    ])
+                    ->where('cliente_origem', $clientName)
+                    ->get()
+                    ->toArray();
+            }
+
             $sql = "SELECT id, nome, username, cnpj, email, telefone, password, data, usuario, celular, cliente_origem 
                     FROM transportadora 
                     WHERE cliente_origem = '" .$clientName."'";  
@@ -26,6 +50,25 @@ class ShippingCompanyRepository{
     public function findAll(){
 
         try{
+            if ($this->canUseEloquent()) {
+                return ShippingCompany::query()
+                    ->select([
+                        'id',
+                        'nome',
+                        'username',
+                        'cnpj',
+                        'email',
+                        'telefone',
+                        'password',
+                        'data',
+                        'usuario',
+                        'celular',
+                        'cliente_origem',
+                    ])
+                    ->get()
+                    ->toArray();
+            }
+
             $sql = "SELECT id, nome, username, cnpj, email, telefone, password, data, usuario, celular, cliente_origem 
                     FROM transportadora";  
                     
@@ -41,6 +84,23 @@ class ShippingCompanyRepository{
     public function save($post){
 
         try{
+            if ($this->canUseEloquent()) {
+                ShippingCompany::query()->create([
+                    'nome' => $post['name'],
+                    'cliente_origem' => 'tetrapak',
+                    'username' => '',
+                    'cnpj' => '',
+                    'email' => '',
+                    'telefone' => '',
+                    'password' => '',
+                    'data' => date('y-m-d H:i:s'),
+                    'usuario' => $_SESSION['nome'],
+                    'celular' => '',
+                ]);
+
+                return 'SAVED';
+            }
+
             $sql = "INSERT INTO transportadora
                     SET nome = '".$post['name']."', cliente_origem = 'tetrapak', username = '', cnpj = '', email = '', telefone = '', password = '', data = '".date("y-m-d H:i:s")."', usuario = '".$_SESSION['nome']."', celular = ''";  
 
@@ -55,6 +115,15 @@ class ShippingCompanyRepository{
     public function updateById($id, $name){
 
         try{
+            if ($this->canUseEloquent()) {
+                ShippingCompany::query()
+                    ->where('id', (int) $id)
+                    ->update([
+                        'nome' => $name,
+                    ]);
+                return 'UPDATED';
+            }
+
             $sql = "UPDATE transportadora
                     SET nome = '".$name."'
                     WHERE ID = ".$id;  
@@ -70,6 +139,13 @@ class ShippingCompanyRepository{
     public function deleteById($id){
 
         try{
+            if ($this->canUseEloquent()) {
+                ShippingCompany::query()
+                    ->where('id', (int) $id)
+                    ->delete();
+                return 'DELETED';
+            }
+
             $sql = "DELETE FROM transportadora
                     WHERE id = ".$id;  
 
@@ -79,6 +155,10 @@ class ShippingCompanyRepository{
         }catch(Exception $e){
             return 'DELETE_ERROR';
         }
+    }
+
+    private function canUseEloquent(){
+        return class_exists(ShippingCompany::class) && class_exists(Capsule::class);
     }
 }
