+++ repository/truckTypeRepository.php
@@ -1,5 +1,9 @@
 
 <?php
+
+use Illuminate\Database\Capsule\Manager as Capsule;
+use Labsoft\Models\TruckType;
+
 class TruckTypeRepository{
 
     private $mySql;
@@ -11,6 +15,13 @@ class TruckTypeRepository{
     public function findAll(){
 
         try{
+            if ($this->canUseEloquent()) {
+                return TruckType::query()
+                    ->select(['id', 'descricao'])
+                    ->get()
+                    ->toArray();
+            }
+
             $sql = "SELECT id, descricao
                     FROM tipoVeiculo";  
 
@@ -26,6 +37,14 @@ class TruckTypeRepository{
     public function findByDescription($description){
 
         try{
+            if ($this->canUseEloquent()) {
+                return TruckType::query()
+                    ->select(['id', 'descricao'])
+                    ->where('descricao', $description)
+                    ->get()
+                    ->toArray();
+            }
+
             $sql = "SELECT id, descricao
                     FROM tipoVeiculo
                     WHERE descricao = '".$description."'";  
@@ -40,6 +59,13 @@ class TruckTypeRepository{
     public function save($post){
 
         try{
+            if ($this->canUseEloquent()) {
+                TruckType::query()->create([
+                    'descricao' => $post['description'],
+                ]);
+                return 'SAVED';
+            }
+
             $sql = "INSERT INTO tipoVeiculo
                     SET descricao = '".$post['description']."'";  
 
@@ -54,6 +80,15 @@ class TruckTypeRepository{
     public function updateById($id, $description){
 
         try{
+            if ($this->canUseEloquent()) {
+                TruckType::query()
+                    ->where('id', (int) $id)
+                    ->update([
+                        'descricao' => $description,
+                    ]);
+                return 'UPDATED';
+            }
+
             $sql = "UPDATE tipoVeiculo
                     SET descricao = '".$description."'
                     WHERE ID = ".$id;  
@@ -69,6 +104,13 @@ class TruckTypeRepository{
     public function deleteById($id){
 
         try{
+            if ($this->canUseEloquent()) {
+                TruckType::query()
+                    ->where('id', (int) $id)
+                    ->delete();
+                return 'DELETED';
+            }
+
             $sql = "DELETE FROM tipoVeiculo
                     WHERE id = ".$id;  
 
@@ -78,6 +120,10 @@ class TruckTypeRepository{
         }catch(Exception $e){
             return 'DELETE_ERROR';
         }
+    }
+
+    private function canUseEloquent(){
+        return class_exists(TruckType::class) && class_exists(Capsule::class);
     }
 }
 ?>
